requirements.txt
Django==1.6.1
awscli==1.3.9
bcdoc==0.12.2
botocore==0.43.0
colorama==0.2.5
dj-database-url==0.2.2
dj-static==0.0.5
django-allauth==0.16.0
django-boto==0.3.1
django-toolbelt==0.0.1
docutils==0.11
gunicorn==18.0
jmespath==0.4.1
oauthlib==0.6.1
pyasn1==0.1.7
pystache==0.5.3
python-dateutil==2.2
python3-openid==3.0.3
requests==2.2.1
requests-oauthlib==0.4.0
rsa==3.1.2
six==1.6.1
static==1.0.2
